Fix convergence and normalization issues in neural-symbolic pipeline
- Fixed decoder output range with Tanh activation for position reconstruction
- Rebalanced initial loss weights to prevent explosion of auxiliary losses
- Implemented reconstruction warmup for first 100 epochs to allow GNN to find basic spatial mapping
- Strengthened spatial clustering with increased collapse prevention weight (1.0 -> 2.0)
- Improved simulator stability by reducing LennardJonesSimulator dt from 0.001 to 0.0005
- Ensured MPS compatibility by casting scatter operations to float32
- Fixed scatter operations to handle None batch tensors properly
- Enhanced loss balancing to prioritize reconstruction during early training

+++ optimized_model.py
@@ -85,8 +85,8 @@ class OptimizedEquivariantGNNLayer(MessagePassing):
     def aggregate(self, inputs, index, dim_size=None):
         # Custom aggregate to handle tuple of (scalar_msg, vector_msg)
         m_h_in, m_v_in = inputs
-        m_h = scatter(m_h_in, index, dim=self.node_dim, dim_size=dim_size, reduce='sum')
-        m_v = scatter(m_v_in, index, dim=self.node_dim, dim_size=dim_size, reduce='sum')
+        m_h = scatter(m_h_in, index, dim=self.node_dim, dim_size=dim_size, reduce='sum').to(torch.float32)
+        m_v = scatter(m_v_in, index, dim=self.node_dim, dim_size=dim_size, reduce='sum').to(torch.float32)
         return m_h, m_v
